lots of small bugs in ipc

+++ NamedPipeIPC/IpcUtils.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.IO;
 using System.Threading;
 using Newtonsoft.Json;
@@ -79,25 +80,66 @@ namespace NamedPipeIPC
 
         }
 
+        public static int WAIT_MILLIS = 100;
+        public static int NUM_RETRIES = 5;
+
+
         public static string SafeReadAllText(string path)
         {
-            using (var fs = new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.ReadWrite))
-            using (var sr = new StreamReader(fs))
+            for (int i = 0; i < NUM_RETRIES; i++)
             {
-                return sr.ReadToEnd();
+                try
+                {
+                    using (var fs = new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.Read))
+                    using (var sr = new StreamReader(fs))
+                    {
+                        return sr.ReadToEnd();
+                    }
+                }
+                catch (Exception e)
+                {
+                    if (i == NUM_RETRIES - 1)
+                    {
+                        throw e;
+                    }
+                    else
+                    {
+                        Thread.Sleep(WAIT_MILLIS);
+                    }
+                }
             }
+            // fail to read
+            return null;
         }
 
         public delegate void DebugLogType(string msg);
 
         public static void SafeWriteAllText(string path, string contents)
         {
-            using (var fs = new FileStream(path, FileMode.Create, FileAccess.Write, FileShare.Read))
-            using (var sw = new StreamWriter(fs))
+            for (int i = 0; i < NUM_RETRIES; i++)
             {
-                sw.Write(contents);
+                try
+                {
+                    using (var fs = new FileStream(path, FileMode.Create, FileAccess.Write, FileShare.None))
+                    using (var sw = new StreamWriter(fs))
+                    {
+                        sw.Write(contents);
+                        return;
+                    }
+                }
+                catch (Exception e)
+                {
+                    if (i == NUM_RETRIES - 1)
+                    {
+                        throw e;
+                    }
+                    else
+                    {
+                        Thread.Sleep(WAIT_MILLIS);
+                    }
+                }
             }
-        }
+         }
 
         /// <summary>
         /// Fetches loaded servers from the config files
@@ -107,14 +149,20 @@ namespace NamedPipeIPC
             List<IpcServerInfo> servers = new List<IpcServerInfo>();
             foreach (string server in Directory.GetFiles(GetServerDirectory(), "*.json")) {
                 try {
-                    IpcServerInfo info = JsonConvert.DeserializeObject<IpcServerInfo>(SafeReadAllText(server));
-                    long timeSinceUpdated = TimeMillis() - info.timeOfLastUpdate;
-                    if  (timeSinceUpdated < keepAliveMillis) {
-                        servers.Add(info);
-                    }
-                    // cleanup very old ones so directory not cluttered
-                    else if (timeSinceUpdated > keepAliveMillis * 4) {
-                        File.Delete(server);
+                    string text = SafeReadAllText(server);
+                    if (text != null)
+                    {
+                        IpcServerInfo info = JsonConvert.DeserializeObject<IpcServerInfo>(text);
+                        long timeSinceUpdated = TimeMillis() - info.timeOfLastUpdate;
+                        if (timeSinceUpdated < keepAliveMillis)
+                        {
+                            servers.Add(info);
+                        }
+                        // cleanup very old ones so directory not cluttered
+                        else if (timeSinceUpdated > keepAliveMillis * 4)
+                        {
+                            File.Delete(server);
+                        }
                     }
                 }
                 catch (JsonSerializationException e) {
